Loads RData and writes to csv
This is an R Markdown version of the R script we were given to load in the data. I also added chunks to create a new folder for each dataset and write each image matrix to a csv within the correct folder. (Sorry the code is ugly, I'm bad at R :)).